#0: move program files into separate folder
