+++ tt_eager/tt_dnn/op_library/operation.hpp
@@ -9,7 +9,7 @@
 #include <tensor/tensor.hpp>
 
 #include "tt_metal/host_api.hpp"
-#include "tt_metal/impl/program.hpp"
+#include "tt_metal/impl/program/program.hpp"
 #include "tt_stl/concepts.hpp"
 #include "tt_stl/reflection.hpp"
