+++ tt_metal/detail/reports/memory_reporter.cpp
@@ -6,7 +6,7 @@
 #include "tt_metal/detail/reports/report_utils.hpp"
 #include "tt_metal/impl/allocator/allocator.hpp"
 #include "tt_metal/impl/device/device.hpp"
-#include "tt_metal/impl/program.hpp"
+#include "tt_metal/impl/program/program.hpp"
 #include <algorithm>
 
 namespace tt::tt_metal {
