+++ tt_metal/impl/module.mk
@@ -14,7 +14,7 @@ TT_METAL_IMPL_SRCS = \
 	tt_metal/impl/allocator/allocator.cpp \
 	tt_metal/impl/allocator/basic_allocator.cpp \
 	tt_metal/impl/allocator/l1_banking_allocator.cpp \
-	tt_metal/impl/program.cpp \
+	tt_metal/impl/program/program.cpp \
 	tt_metal/impl/dispatch/device_command.cpp \
 	tt_metal/impl/dispatch/debug_tools.cpp \
 	tt_metal/impl/dispatch/command_queue.cpp
